Moving the connection to the car at first page
