+++ Project2/app/src/main/java/com/example/project/Main2Activity.java
@@ -1,36 +1,30 @@
 package com.example.project;
 
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothDevice;
-import android.bluetooth.BluetoothSocket;
-import android.content.Intent;
+
+import android.annotation.SuppressLint;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
-import android.widget.Toast;
+
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.util.Set;
-import java.util.UUID;
+
 
 public class Main2Activity extends AppCompatActivity {
-    
-    private final String DEVICE_ADDRESS= "20:15:10:20:05:64"; //MAC Address of the bluetooth module
-    private static final UUID MY_UUID = UUID.fromString("00001101-0000-1000-8000-00805f9b34fb");
 
-    private BluetoothDevice device;
-    private BluetoothSocket socket;
+
     private OutputStream outputStream;
 
-    Button forward, left, right, reverse, bluetooth_connect_btn;
+    Button forward, left, right, reverse;
 
     String command;
 
 
 
+    @SuppressLint("ClickableViewAccessibility")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -40,7 +34,6 @@ public class Main2Activity extends AppCompatActivity {
         left = (Button) findViewById(R.id.left);
         right = (Button) findViewById(R.id.right);
         reverse = (Button) findViewById(R.id.backwards);
-        bluetooth_connect_btn = (Button) findViewById(R.id.bluetooth_connect_btn);
 
         forward.setOnTouchListener(new View.OnTouchListener() {
             @Override
@@ -49,7 +42,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "F";
                     try {
-                        outputStream.write(command.toString().getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -57,7 +50,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "0";
                     try {
-                        outputStream.write(command.getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -72,7 +65,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "B";
                     try {
-                        outputStream.write(command.toString().getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -80,7 +73,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "0";
                     try {
-                        outputStream.write(command.getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -96,7 +89,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "L";
                     try {
-                        outputStream.write(command.toString().getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -104,7 +97,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "0";
                     try {
-                        outputStream.write(command.getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -119,7 +112,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "R";
                     try {
-                        outputStream.write(command.toString().getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -127,7 +120,7 @@ public class Main2Activity extends AppCompatActivity {
                 {
                     command = "0";
                     try {
-                        outputStream.write(command.getBytes());
+                        MainActivity.btSocket.getOutputStream().write(command.getBytes());
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
@@ -135,85 +128,8 @@ public class Main2Activity extends AppCompatActivity {
                 return false;
             }
         });
-        bluetooth_connect_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                if (BTinitialize()) {
-                    BTconnect();
-                }
-
-            }
-        });
     }
-        public boolean BTinitialize()
-        {
-            boolean found = false;
-            BluetoothAdapter bluetoothAdapter = BluetoothAdapter.getDefaultAdapter();
-            if(bluetoothAdapter == null)
-            {
-                Toast.makeText(getApplicationContext(), "Device doesn't support bluetooth", Toast.LENGTH_SHORT).show();
-            }
-            if(!bluetoothAdapter.isEnabled())
-            {
-                Intent enableAdapter = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
-                startActivityForResult(enableAdapter,0);
-                try
-                {
-                    Thread.sleep(1000);
-                }
-                catch(InterruptedException e)
-                {
-                    e.printStackTrace();
-                }
-            }
-            Set<BluetoothDevice> bondedDevices = bluetoothAdapter.getBondedDevices();
-            if(bondedDevices.isEmpty())
-            {
-                Toast.makeText(getApplicationContext(), "Please pair the device first", Toast.LENGTH_SHORT).show();
-            }
-            else
-            {
-                for(BluetoothDevice iterator : bondedDevices)
-                {
-                    if(iterator.getAddress().equals(DEVICE_ADDRESS))
-                    {
-                        device = iterator;
-                        found = true;
-                        break;
-                    }
-                }
-            }
-            return found;
-        }
-        public boolean BTconnect()
-        {
-            boolean connected = true;
-            try
-            {
-                socket = device.createRfcommSocketToServiceRecord(MY_UUID);
-                socket.connect();
-                Toast.makeText(getApplicationContext(),
-                        "Connection to bluetooth device successful", Toast.LENGTH_LONG).show();
-            }
-            catch(IOException e)
-            {
-                e.printStackTrace();
-                connected = false;
-            }
-            if(connected)
-            {
-                try
-                {
-                    outputStream = socket.getOutputStream();
-                }
-                catch(IOException e)
-                {
-                    e.printStackTrace();
-                }
-            }
-            return connected;
-        }
+
         @Override
         protected void onStart()
         {
