+++ Project2/app/src/main/java/com/example/project/Parking.java
@@ -2,31 +2,19 @@ package com.example.project;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.bluetooth.BluetoothSocket;
-import android.content.Intent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 import android.app.ProgressDialog;
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothDevice;
-import android.os.AsyncTask;
 import java.io.IOException;
-import java.util.UUID;
 
 
 public class Parking extends AppCompatActivity {
 
-    Button connect;
     Button park;
 
-    final String  address = "20:15:10:20:05:64"; ;
     private ProgressDialog progress;
-    BluetoothAdapter myBluetooth = null;
-    BluetoothSocket btSocket = null;
-    private boolean isBtConnected = false;
 
-    static final UUID myUUID = UUID.fromString("00001101-0000-1000-8000-00805F9B34FB");
 
     @Override
     protected void onCreate(Bundle savedInstanceState)
@@ -36,7 +24,6 @@ public class Parking extends AppCompatActivity {
         setContentView(R.layout.activity_parking);
 
         //call the widgets
-        connect = (Button)findViewById(R.id.connect);
 
         park = (Button) findViewById(R.id.park);
 
@@ -44,42 +31,16 @@ public class Parking extends AppCompatActivity {
          //Call the class to connect
 
         //commands to be sent to bluetooth
-        connect.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View v)
-            {
-                new ConnectBT().execute();
-            }
-        });
+
 
         park.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if (btSocket!=null)
+                if (MainActivity.btSocket!=null)
                 {
                     try
                     {
-                        btSocket.getOutputStream().write("P".toString().getBytes());
-                    }
-                    catch (IOException e)
-                    {
-                        msg("Error");
-                    }
-                }
-            }
-        });
-
-        Button cruise = (Button) findViewById(R.id.cruise);
-
-        cruise.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (btSocket!=null)
-                {
-                    try
-                    {
-                        btSocket.getOutputStream().write("C".toString().getBytes());
+                        MainActivity.btSocket.getOutputStream().write("P".getBytes());
                     }
                     catch (IOException e)
                     {
@@ -94,56 +55,5 @@ public class Parking extends AppCompatActivity {
     private void msg(String s)
     {
         Toast.makeText(getApplicationContext(),s,Toast.LENGTH_LONG).show();
-    }
-
-
-
-    private class ConnectBT extends AsyncTask<Void, Void, Void>  // UI thread
-    {
-        private boolean ConnectSuccess = true; //if it's here, it's almost connected
-
-        @Override
-        protected void onPreExecute()
-        {
-            progress = ProgressDialog.show(Parking.this, "Connecting...", "Please wait!!!");  //show a progress dialog
-        }
-
-        @Override
-        protected Void doInBackground(Void... devices) //while the progress dialog is shown, the connection is done in background
-        {
-            try
-            {
-                if (btSocket == null || !isBtConnected)
-                {
-                    myBluetooth = BluetoothAdapter.getDefaultAdapter();//get the mobile bluetooth device
-                    BluetoothDevice dispositivo = myBluetooth.getRemoteDevice(address);//connects to the device's address and checks if it's available
-                    btSocket = dispositivo.createInsecureRfcommSocketToServiceRecord(myUUID);//create a RFCOMM (SPP) connection
-                    BluetoothAdapter.getDefaultAdapter().cancelDiscovery();
-                    btSocket.connect();//start connection
-                }
-            }
-            catch (IOException e)
-            {
-                ConnectSuccess = false;//if the try failed, you can check the exception here
-            }
-            return null;
-        }
-        @Override
-        protected void onPostExecute(Void result) //after the doInBackground, it checks if everything went fine
-        {
-            super.onPostExecute(result);
-
-            if (!ConnectSuccess)
-            {
-                msg("Connection Failed. Is it a SPP Bluetooth? Try again.");
-                finish();
-            }
-            else
-            {
-                msg("Connected.");
-                isBtConnected = true;
-            }
-            progress.dismiss();
-        }
     }
 }
